+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added tag disabled examples for `terra-form-multi-select` and `terra-form-multi-select-field`.
 
+* Updated
+  * Updated Draggable list item example for selection state in `terra-list`.
+
 ## 1.56.0 - (January 9, 2024)
 
 * Changed